Update order closing functionality and UI feedback
In this commit, the order closing functionality was significantly updated. The button text in `AddNewOrderControl.xaml` was changed and now triggers the new `CloseAndSaveOrderCommand`. This command was added to the `AddNewOrderViewModel` class and is initialized in its constructor to call the `CloseAndSaveOrder` method. This new method asks the user for confirmation twice before closing and saving an order. Additionally, several message strings were updated for better user feedback. A `Refresh` method was added to `OrderControl.xaml.cs` to refresh the control content. In `OrderRepository.cs`, a `CloseOrder` method was added to handle order closing in the database. Lastly, the exit confirmation dialog in `MainWindow.xaml.cs` was changed from a `MessageBox` to a `MessageBoxWindow`.