UI redesign in AddNewOrderControl.xaml
Redesigned the user interface of AddNewOrderControl.xaml to enhance usability and visual appeal. Adjusted the overall design size, increasing height from 850 to 950 and width from 1300 to 1800, to better accommodate content and improve readability. Reorganized elements for a more intuitive user experience: moved and aligned the WrapPanel containing the item number input to the top and left, respectively, and integrated the "Keresés" button into this panel for cohesion. Adjusted the button's size and font for better interaction. Changed the product description background to "Transparent" for a seamless UI integration.