Update ListViewItem background and add DataTrigger
This commit updates the default background color of ListViewItem elements from `#F4F4F4` to `LightGray`. Additionally, a new DataTrigger has been introduced to the ListViewItem style, which sets the background color back to `#F4F4F4` when the `IsOpen` property in the data binding context is set to `True`.