Refactor OrderDetailsViewModel Order property
The `Order` property in the `OrderDetailsViewModel` class has been updated to use a private backing field `_order` instead of being an auto-implemented property. This change introduces a getter that returns the private field and a setter that assigns the value to the private field and calls `OnPropertyChanged()` to notify observers of the property change. This enhancement improves encapsulation and supports additional logic upon setting the property, such as UI updates.