Add ShowAllNoteControl UserControl and ViewModel
Introduced a new UserControl named `ShowAllNoteControl` in the
`Ol_der.Controls.Notes` namespace. This includes the addition of
`ShowAllNoteControl.xaml` and its code-behind file
`ShowAllNoteControl.xaml.cs`. The XAML file defines a basic structure
with a `Grid` element and necessary XML namespaces. The code-behind
file includes the class definition and a constructor to initialize
the component. Additionally, added an empty `ShowAllNoteViewModel`
class in `ShowAllNoteViewModel.cs`, marked as `internal`.