Add AddNewNoteViewModel with save functionality
Added a new ViewModel class `AddNewNoteViewModel` to the `Ol_der.Controls.Notes` namespace. Implemented the `INotifyPropertyChanged` interface to support property change notifications. Introduced a private field `_note` and a public property `Note` to hold the note data. Added a `SaveNoteCommand` property of type `ICommand` to handle the save note action. Created a constructor that initializes the `NoteRepository`, sets up the `SaveNoteCommand` with a `RelayCommand`, and initializes the `Note` property. Implemented the `SaveNote` method to set the creation date of the note, save it using the `NoteRepository`, and display a confirmation message box. Added the `OnPropertyChanged` method to raise the `PropertyChanged` event when a property value changes.