Add "Jegyzetek" button and handle its click event
A new "Jegyzetek" button has been added to the MainWindow.xaml file.
The MainWindow.xaml.cs file has been updated to include a new using
directive for the Ol_der.Controls.Notes namespace. Additionally, a
new case for the "btnNote" button has been added to the
Menu_Button_Click event handler to display a NoteControl in the
ContentArea when the button is clicked.