Increase font sizes and adjust TextBox dimensions
Added a `FontSize` attribute to the root element with a value of `18`.
Increased the `FontSize` of the `Label` in the first row from `20` to `22`.
Adjusted the `Width` of the `TextBox` in the second row from `500` to `600`.
Adjusted the `Height` of the `TextBox` in the second row from `100` to `125`.
Increased the `FontSize` of the `Button` in the third row from `20` to `22`.