Add Note parameter to AddNewNoteControl constructor
The constructor of the `AddNewNoteControl` class has been modified to accept a `Note` object as a parameter. This allows the control to be initialized with an existing note, which is then passed to the `AddNewNoteViewModel`. The `_viewModel` field is now initialized with the `NoteToUpdate` parameter, enabling the view model to work with the provided note. Additionally, a new using directive for `Ol_der.Models` has been added to support the use of the `Note` class.