Add OnOrderFinished event to AddNewOrderControl
Introduce an Action delegate named OnOrderFinished to the AddNewOrderControl class. Update the constructor to manage the subscription to this event. Add a new method, ActivateEvent, to invoke the OnOrderFinished action if it is not null. This change allows external code to subscribe to the OnOrderFinished event and perform actions when the event is triggered.