Add note detail view functionality to NoteControl
Added a private field `_noteDetailControl` of type `NoteDetailControl` to the `NoteControl` class. Introduced a new method `Note_Detail_Click` to handle the click event for viewing note details. The method retrieves the selected note using `_showAllNoteControl.GetSelectedNote()`. If no note is selected, a message box with the text "Nincs kiválasztott jegyzet a nagy nézethez!" is displayed. If a note is selected, a new `NoteDetailControl` is instantiated with the selected note and set as the content of `ContentArea`.