Increase WrapPanel width and Border margin in ListBox
The width of the `WrapPanel` inside the `ItemsPanelTemplate` of the
`ListBox` has been increased from 1280 to 1500. This adjustment
provides more space for the items within the `ListBox`, potentially
accommodating more items in a single row or providing a better layout.

Additionally, the `Margin` of the `Border` inside the `DataTemplate`
of the `ListBox` has been increased from 10 to 30. This change
introduces more spacing between the items, improving readability
and visual appeal.