Add WarrantyStatuses property to Warranty class
The Warranty class in the Ol_der.Models namespace now includes a new property, WarrantyStatuses, of type ICollection<WarrantyStatus>. This property is initialized to a new List<WarrantyStatus>, allowing the Warranty class to maintain a collection of WarrantyStatus objects. This enhancement facilitates tracking the status history or various states of a warranty.