Add ShowAllWarrantyControl UserControl and ViewModel
Added a new UserControl named `ShowAllWarrantyControl` in the
`ShowAllWarrantyControl.xaml` file. This control includes a `ListView`
to display warranty information with columns for ID, Supplier, Product
Name, Creation Date, and Completion Status. The `ListView` also includes
styling to change the background color of completed warranties.

Created a code-behind file `ShowAllWarrantyControl.xaml.cs` for the
`ShowAllWarrantyControl` UserControl. This file includes necessary using
directives and a constructor that initializes the component.

Added a new ViewModel class named `ShowAllWarrantyViewModel` in the
`ShowAllWarrantyViewModel.cs` file. This class is currently empty and
marked as internal.