Add CustomerOrderRepository with limited order retrieval
A new `CustomerOrderRepository` class has been added to the
`Ol_der.Controls.CustomerOrders` namespace. This class includes
a method `GetLimitedNumberOfCustomerOrderAsync` which retrieves
a limited number of customer orders from the database using
Entity Framework Core. The method queries the `CustomerOrders`
table, includes related `CustomerOrderStatuses`, and orders the
results by `CreationDate` in descending order. It takes an
integer parameter `limit` to specify the maximum number of
customer orders to retrieve. The `ApplicationDbContextFactory.
Create()` method is used to create a new instance of the
database context.