Refactor CustomerOrderDetailControl to use ViewModel
Added `using` directive for `Ol_der.Models`. Modified the constructor of `CustomerOrderDetailControl` to accept a `CustomerOrder` object. Introduced a private `_viewModel` field of type `CustomerOrderDetailViewModel` and initialized it with the `customerOrderDetails` parameter. Set the `DataContext` of the control to the `_viewModel`. These changes enhance the control's ability to display and manage customer order details using the MVVM pattern.