Refactor and restyle ListView DataTemplate in XAML
Significantly restructured and updated the styling of the
DataTemplate in `ShowAllPackageControl.xaml`. Replaced the
previous `StackPanel` containing a `Grid` and an `ItemsControl`
with a `Border` element featuring a `BorderBrush` of `#dfd3c3`
and a `BorderThickness` of `2`. The new structure includes a
`StackPanel` with a background color of `#dfd3c3` that contains
the `Grid` for displaying the main data. The `ItemsControl` for
displaying `SaleItems` is now nested within this new `Border`
structure, enhancing the visual appearance and organization of
the items within the `ListView`.