Refactor ComboBox styling in SelectSupplierWindow
Updated the `Window.Resources` to include a modern ComboBox style from `/Styles/ModernComboBoxStyles.xaml`. Removed the custom `Style` definition for the `ComboBox` and applied the new `Style` using `{StaticResource ModernComboBoxForSuppliers}`.

Added a `TextBlock` above the `ComboBox` to display a label prompting the user to select a supplier. Minor adjustments were made to the `ComboBox` properties to maintain functionality while adopting the new style.